fixing rpc server for dns reports. (#874)
* fixing rpc server for dns reports.

* fix: mac builds
