+++ controller/internal/enforcer/dnsproxy/dns_test.go
@@ -1,3 +1,5 @@
+// +build linux !darwin
+
 package dnsproxy
 
 import (
